+++ src/lib/components/ItemCard.svelte
@@ -21,9 +21,6 @@
 		locationsLoaded = true;
 	});
 
-	console.log($AllLocations);
-	console.log(item.data.location);
-
 	$effect(() => {
 		if (locationsLoaded && item.data?.location) {
 			location = $AllLocations.find((loc) => loc.id === item.data.location);
@@ -72,7 +69,7 @@
 		{/if}
 
 		<div id="cardContent" class="justify-between py-4 md:flex md:gap-6 items-center">
-			<div class="mb-5 h-52 w-full overflow-hidden rounded-2xl md:aspect-[3/2] md:w-auto">
+			<div class="mb-5 h-52 w-full overflow-hidden rounded-2xl md:aspect-3/2 md:w-auto">
 				<img class="h-full w-full object-cover" src={item.data.image} alt={item.data.imageAlt} loading="lazy" />
 			</div>
 			<div id="cardText" class="text-md w-full md:w-[60%]">